refactor(circle-loader): matching loader sibling naming
