+++ packages/circle-loader/sp-circle-loader.ts
@@ -9,10 +9,12 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import '@spectrum-web-components/circleloader/sp-circleloader.js';
-import { html } from 'lit-html';
-import { measureFixtureCreation } from '../../../../test/benchmark/helpers';
+import { CircleLoader } from './src/CircleLoader.js';
 
-measureFixtureCreation(html`
-    <sp-circleloader indeterminate></sp-circleloader>
-`);
+customElements.define('sp-circle-loader', CircleLoader);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'sp-circle-loader': CircleLoader;
+    }
+}
