+++ packages/circle-loader/src/CircleLoader.ts
