+++ packages/circle-loader/src/spectrum-config.js
