+++ packages/circle-loader/test/benchmark/test-basic.ts
@@ -9,12 +9,10 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { CircleLoader } from './src/CircleLoader.js';
+import '@spectrum-web-components/circleloader/sp-circle-loader.js';
+import { html } from 'lit-html';
+import { measureFixtureCreation } from '../../../../test/benchmark/helpers';
 
-customElements.define('sp-circleloader', CircleLoader);
-
-declare global {
-    interface HTMLElementTagNameMap {
-        'sp-circleloader': CircleLoader;
-    }
-}
+measureFixtureCreation(html`
+    <sp-circleloader indeterminate></sp-circleloader>
+`);
